services/api/requirements.txt

fastapi==0.115.2
uvicorn[standard]==0.30.6
motor==3.6.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
slowapi==0.1.9
oqs==0.10.0
orjson==3.10.7
python-multipart==0.0.12
